# Conflicts:
#	app/src/main/res/layout/activity_login.xml
#	app/src/main/res/layout/activity_register.xml

+++ app/src/main/java/com/phule/mtstudentinformationmanagement/StudentAdapter.java
@@ -3,6 +3,7 @@ package com.phule.mtstudentinformationmanagement;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
@@ -14,17 +15,25 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import org.w3c.dom.Text;
 
 import java.util.List;
 
 public class StudentAdapter extends RecyclerView.Adapter<StudentAdapter.ViewHolder> {
+    private static StudentAdapter instance;
     private List<Student> studentList;
     public StudentAdapter(List<Student> studentList) {
 
         this.studentList = studentList;
+        instance = this;
+    }
+    public static StudentAdapter getInstance() {
+        return instance;
     }
     @NonNull
     @Override
@@ -48,8 +57,19 @@ public class StudentAdapter extends RecyclerView.Adapter<StudentAdapter.ViewHold
                     public boolean onMenuItemClick(MenuItem menuItem) {
                         if(menuItem.getItemId() == R.id.menu_edit) {
                             // Edit action
+                            Intent intent = new Intent(view.getContext(), EditStudentActivity.class);
+                            intent.putExtra("code", student.getCode());
+                            intent.putExtra("name", student.getName());
+                            intent.putExtra("birthday", student.getBirthday());
+                            intent.putExtra("address", student.getAddress());
+                            intent.putExtra("gender", student.getGender());
+                            intent.putExtra("phone", student.getPhone());
+                            intent.putExtra("enrollmentDate", student.getEnrollmentDate());
+                            intent.putExtra("major", student.getMajor());
+                            view.getContext().startActivity(intent);
                         }
                         else if(menuItem.getItemId() == R.id.menu_remove) {
+                            // Remove action
                             new AlertDialog.Builder(view.getContext())
                                     .setTitle(view.getContext().getString(R.string.confirm_removal_title))
                                     .setMessage(view.getContext().getString(R.string.confirm_removal_msg))
@@ -58,23 +78,44 @@ public class StudentAdapter extends RecyclerView.Adapter<StudentAdapter.ViewHold
                                         public void onClick(DialogInterface dialog, int whichButton) {
                                             // Remove student from Firestore
                                             FirebaseFirestore db = FirebaseFirestore.getInstance();
-                                            db.collection("Students").document(student.getDocumentId())
-                                                    .delete()
-                                                    .addOnSuccessListener(aVoid -> {
-                                                        // Remove student from studentList recycler view
-                                                        studentList.remove(position);
-                                                        notifyItemRemoved(position);
-                                                        Toast.makeText(view.getContext(), "Student removed", Toast.LENGTH_SHORT).show();
-                                                    })
-                                                    .addOnFailureListener(e -> {
-                                                        // Handle Firestore deletion failure
-                                                        Toast.makeText(view.getContext(), "Error removing student", Toast.LENGTH_SHORT).show();
+                                            db.collection("Students").whereEqualTo("code", student.getCode()).get()
+                                                    .addOnCompleteListener(task -> {
+                                                        if (task.isSuccessful()) {
+                                                            if (task.getResult().isEmpty()) {
+                                                                Log.d("removeStudent", "No matching documents found");
+                                                                Toast.makeText(view.getContext(), "No matching documents found", Toast.LENGTH_SHORT).show();
+                                                                return;
+                                                            }
+                                                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                                                String documentId = document.getId();
+
+                                                                db.collection("Students").document(documentId)
+                                                                        .delete()
+                                                                        .addOnSuccessListener(aVoid -> {
+                                                                            int position = holder.getAdapterPosition();
+                                                                            studentList.remove(position);
+                                                                            notifyItemRemoved(position);
+                                                                            Toast.makeText(view.getContext(), "Student removed", Toast.LENGTH_SHORT).show();
+                                                                            Log.d("removeStudent", "Student removed successfully: " + student.getCode());
+                                                                        })
+                                                                        .addOnFailureListener(e -> {
+                                                                            // Handle Firestore deletion failure
+                                                                            Toast.makeText(view.getContext(), "Error removing student", Toast.LENGTH_SHORT).show();
+                                                                            Log.d("removeStudent", "Student removed failed: " + student.getCode());
+                                                                        });
+                                                            }
+                                                        } else {
+                                                            Log.d("removeStudent", "Failed to find code");
+                                                            Toast.makeText(view.getContext(), "Failed to find student code", Toast.LENGTH_SHORT).show();
+                                                        }
+                                                    }).addOnFailureListener(e -> {
+                                                        Toast.makeText(view.getContext(), "Error fetching documents", Toast.LENGTH_SHORT).show();
+                                                        Log.d("removeStudent", "Failed to fetch documents: " + e.getMessage());
                                                     });
                                         }
                                     })
                                     .setNegativeButton(view.getContext().getString(R.string.confirm_removal_deny), null)
                                     .show();
-
                         }
                         return true;
                     }
@@ -83,15 +124,7 @@ public class StudentAdapter extends RecyclerView.Adapter<StudentAdapter.ViewHold
                 popupMenu.show();
             }
         });
-    }
 
-    private void removeStudentById(String studentId) {
-        for (int i = 0; i < studentList.size(); i++) {
-            if (studentList.get(i).getDocumentId().equals(studentId)) {
-                studentList.remove(i);
-                return;
-            }
-        }
     }
     @Override
     public int getItemCount() {
